restructure model loader
